Split the metadata handler into a separate file.
This is only a cleanup as a prep for future changes.
